requirements.txt
streamlit
transformers 
sentence-transformers 
scikit-learn
torch 
faiss-cpu 
rapidfuzz 
scikit-learn 
plotly 
openai 
google-generativeai
vertexai